requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
ping3
plotly
pandas